admin restrictions to edit and delete restaurant

+++ client/src/components/restaurants/Restaurant.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { Card, Button, Icon, } from "semantic-ui-react";
 import RestaurantForm from './RestaurantForm'
+import { AuthConsumer, } from "../../providers/AuthProvider";
+import { withRouter, } from 'react-router-dom';
 
 class Restaurant extends React.Component {
   state = {restaurant: {editing: false, }, }
@@ -41,35 +43,61 @@ class Restaurant extends React.Component {
       )
     }
 
+  adminCheck = () => {
+    const { restaurant, auth: { user, } } = this.props
+
+    if (user.admin == true) {
+      return (
+        <>
+          { this.state.editing ? this.restaurantEdit() : this.restaurantView() }
+          <Button
+            icon
+            color="blue"
+            size="tiny"
+            onClick={() => this.toggleEdit(this.props.id)}
+          >
+            { this.state.editing ? 'Cancel'
+            :
+            <Icon name="pencil" />
+            }
+          </Button>
+          <Button
+            icon
+            color="red"
+            size="tiny"
+            onClick={ () => this.props.deleteRestaurant(this.props.id) }
+            style={{ marginLeft: "15px", }}
+          >
+            <Icon name ="trash" />
+          </Button>
+        </>
+      )
+    } else {
+      return(
+        this.restaurantView() 
+      )
+    }
+  }
+
   render() {
-    return (
-      <>
-        { this.state.editing ? this.restaurantEdit() : this.restaurantView() }
-        <Button
-          icon
-          color="blue"
-          size="tiny"
-          onClick={() => this.toggleEdit(this.props.id)}
-        >
-          { this.state.editing ? 'Cancel'
-          :
-          <Icon name="pencil" />
-          }
-        </Button>
-        <Button
-          icon
-          color="red"
-          size="tiny"
-          onClick={ () => this.props.deleteRestaurant(this.props.id) }
-          style={{ marginLeft: "15px", }}
-        >
-          <Icon name ="trash" />
-        </Button>
-      </>
+    return(
+      <div>
+        { this.adminCheck() }
+      </div>
     )
   }
 }
 
-  //onclick of blue pencil want to render form where can edit!
+export class ConnectedRestaurant extends React.Component {
+  render() {
+    return(
+      <AuthConsumer>
+        { auth =>
+            <Restaurant {... this.props } {...this.state} auth={auth} />
+        }
+      </AuthConsumer>
+    )
+  }
+}
 
-export default Restaurant;
+export default withRouter(ConnectedRestaurant);
